% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.lapply-package.R
\docType{package}
\name{proj.lapply}
\alias{proj.lapply-package}
\alias{proj.lapply}
\title{\code{\link{proj.lapply}}}
\description{
Scientific project development and management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edwardlavender/proj.lapply/}
  \item Report bugs at \url{https://github.com/edwardlavender/proj.lapplys/issues/}
}

}
\author{
\strong{Maintainer}: Edward Lavender \email{edward.lavender@eawag.ch} (\href{https://orcid.org/0000-0002-8040-7489}{ORCID})

}
