% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow.R
\name{workflow}
\alias{workflow}
\title{Workflows}
\usage{
workflow(
  .sim,
  .datasets,
  .constructor,
  ...,
  .algorithm,
  .trials,
  .success,
  .write,
  .coffee,
  .verbose
)
}
\arguments{
\item{.sim, .datasets, .constructor, .algorithm, .trials, .success, .write, .coffee}{Arguments directly inherited from \code{\link[=cl_lapply_workflow]{cl_lapply_workflow()}}. \code{sim} is the \code{iteration} \code{\link[data.table:data.table]{data.table::data.table}} row.}

\item{.verbose}{A \code{logical} variable that defines whether or not to send user outputs to the console or a \code{.txt} file.}
}
\description{
This internal function is the workhorse of \code{\link[=cl_lapply_workflow]{cl_lapply_workflow()}}. The function wraps an \code{algorithm} function.
}
\details{
For details, see the wrapper function \code{\link[=cl_lapply_workflow]{cl_lapply_workflow()}}.
}
\author{
Edward Lavender
}
\keyword{internal}
